+++ pkg/crmanager/worker.go
@@ -73,8 +73,8 @@ func (crMgr *CRManager) processResource() bool {
 		if crMgr.initState {
 			break
 		}
-		tls := rKey.rsc.(*cisapiv1.TLSProfile)
-		virtuals := crMgr.getVirtualsForTLSProfile(tls)
+		tlsProfile := rKey.rsc.(*cisapiv1.TLSProfile)
+		virtuals := crMgr.getVirtualsForTLSProfile(tlsProfile)
 		// No Virtuals are effected with the change in TLSProfile.
 		if nil == virtuals {
 			break
@@ -112,11 +112,19 @@ func (crMgr *CRManager) processResource() bool {
 		ipam := rKey.rsc.(*ficV1.F5IPAM)
 		virtuals := crMgr.getVirtualServersForIPAM(ipam)
 		for _, vs := range virtuals {
-			crMgr.processVirtualServers(vs, false)
+			err := crMgr.processVirtualServers(vs, false)
+			if err != nil {
+				utilruntime.HandleError(fmt.Errorf("Sync %v failed with %v", key, err))
+				isError = true
+			}
 		}
 		TSVirtuals := crMgr.getTransportServersForIPAM(ipam)
 		for _, ts := range TSVirtuals {
-			crMgr.processTransportServers(ts, false)
+			err := crMgr.processTransportServers(ts, false)
+			if err != nil {
+				utilruntime.HandleError(fmt.Errorf("Sync %v failed with %v", key, err))
+				isError = true
+			}
 		}
 
 		services := crMgr.syncAndGetServicesForIPAM(ipam)
@@ -396,19 +404,24 @@ func (crMgr *CRManager) getAllVirtualServers(namespace string) []*cisapiv1.Virtu
 		log.Errorf("Informer not found for namespace: %v", namespace)
 		return nil
 	}
-	// Get list of VirtualServers and process them.
-	orderedVSs, err := crInf.vsInformer.GetIndexer().ByIndex("namespace", namespace)
-	if err != nil {
-		log.Errorf("Unable to get list of VirtualServers for namespace '%v': %v",
-			namespace, err)
-		return nil
+
+	var orderedVSs []interface{}
+	var err error
+	if namespace == "" {
+		orderedVSs = crInf.vsInformer.GetIndexer().List()
+	} else {
+		// Get list of VirtualServers and process them.
+		orderedVSs, err = crInf.vsInformer.GetIndexer().ByIndex("namespace", namespace)
+		if err != nil {
+			log.Errorf("Unable to get list of VirtualServers for namespace '%v': %v",
+				namespace, err)
+			return nil
+		}
 	}
 
 	for _, obj := range orderedVSs {
 		vs := obj.(*cisapiv1.VirtualServer)
-		// TODO
-		// Validate the VirtualServers List to check if all the vs are valid.
-
+		// TODO: Validate the VirtualServers List to check if all the vs are valid.
 		allVirtuals = append(allVirtuals, vs)
 	}
 
@@ -416,23 +429,13 @@ func (crMgr *CRManager) getAllVirtualServers(namespace string) []*cisapiv1.Virtu
 }
 
 // getAllVirtualServers returns list of all valid VirtualServers in rkey namespace.
-func (crMgr *CRManager) getAllVSFromAllNamespaces() []*cisapiv1.VirtualServer {
+func (crMgr *CRManager) getAllVSFromMonitoredNamespaces() []*cisapiv1.VirtualServer {
 	var allVirtuals []*cisapiv1.VirtualServer
-
-	crInf, ok := crMgr.getNamespacedInformer("")
-	if !ok {
-		log.Errorf("Informer not found all namespace.")
-		return allVirtuals
+	if crMgr.watchingAllNamespaces() {
+		return crMgr.getAllVirtualServers("")
 	}
-	// Get list of VirtualServers and process them.
-	objs := crInf.vsInformer.GetIndexer().List()
-
-	for _, obj := range objs {
-		vs := obj.(*cisapiv1.VirtualServer)
-		// TODO
-		// Validate the VirtualServers List to check if all the vs are valid.
-
-		allVirtuals = append(allVirtuals, vs)
+	for ns := range crMgr.namespaces {
+		allVirtuals = append(allVirtuals, crMgr.getAllVirtualServers(ns)...)
 	}
 
 	return allVirtuals
@@ -654,11 +657,11 @@ func (crMgr *CRManager) processVirtualServers(
 			ip = virtual.Spec.VirtualServerAddress
 		} else {
 			ip = crMgr.requestIP(ipamLabel, virtual.Spec.Host, "")
-			log.Debugf("[ipam] requested IP for host %v is: %v", virtual.Spec.Host, ip)
 			if ip == "" {
 				log.Debugf("[ipam] requested IP for host %v is empty.", virtual.Spec.Host)
 				return nil
 			}
+			log.Debugf("[ipam] requested IP for host %v is: %v", virtual.Spec.Host, ip)
 		}
 	} else {
 		if virtual.Spec.VirtualServerAddress == "" {
@@ -771,12 +774,12 @@ func (crMgr *CRManager) processVirtualServers(
 func (crMgr *CRManager) getIPAMCR() *ficV1.F5IPAM {
 	cr := strings.Split(crMgr.ipamCR, "/")
 	if len(cr) != 2 {
-		log.Errorf("[ipam] error while retriving IPAM namespace and name.")
+		log.Errorf("[ipam] error while retrieving IPAM namespace and name.")
 		return nil
 	}
 	ipamCR, err := crMgr.ipamCli.Get(cr[0], cr[1])
 	if err != nil {
-		log.Errorf("[ipam] error while retriving IPAM custom resource.")
+		log.Errorf("[ipam] error while retrieving IPAM custom resource.")
 		return nil
 	}
 	return ipamCR
@@ -849,8 +852,12 @@ func (crMgr *CRManager) requestIP(ipamLabel string, host string, key string) str
 		return ""
 	}
 
-	crMgr.ipamCli.Update(IPAMNamespace, ipamCR)
-	log.Debugf("[ipam] Updated IPAM CR.")
+	_, err := crMgr.ipamCli.Update(IPAMNamespace, ipamCR)
+	if err != nil {
+		log.Errorf("[ipam] Error updating IPAM CR : %v", err)
+	} else {
+		log.Debugf("[ipam] Updated IPAM CR.")
+	}
 	return ""
 
 }
@@ -1212,24 +1219,14 @@ func (crMgr *CRManager) processTransportServers(
 }
 
 // getAllTransportServers returns list of all valid TransportServers in rkey namespace.
-func (crMgr *CRManager) getAllTSFromAllNamespaces() []*cisapiv1.TransportServer {
+func (crMgr *CRManager) getAllTSFromMonitoredNamespaces() []*cisapiv1.TransportServer {
 	var allVirtuals []*cisapiv1.TransportServer
-
-	crInf, ok := crMgr.getNamespacedInformer("")
-	if !ok {
-		log.Errorf("Informer not found for all namespace.")
-		return allVirtuals
+	if crMgr.watchingAllNamespaces() {
+		return crMgr.getAllTransportServers("")
 	}
-	// Get list of VirtualServers and process them.
-	orderedVSs := crInf.tsInformer.GetIndexer().List()
-	for _, obj := range orderedVSs {
-		vs := obj.(*cisapiv1.TransportServer)
-		// TODO
-		// Validate the TransportServers List to check if all the vs are valid.
-
-		allVirtuals = append(allVirtuals, vs)
+	for ns := range crMgr.namespaces {
+		allVirtuals = append(allVirtuals, crMgr.getAllTransportServers(ns)...)
 	}
-
 	return allVirtuals
 }
 
@@ -1242,19 +1239,22 @@ func (crMgr *CRManager) getAllTransportServers(namespace string) []*cisapiv1.Tra
 		log.Errorf("Informer not found for namespace: %v", namespace)
 		return nil
 	}
-	// Get list of VirtualServers and process them.
-	orderedVSs, err := crInf.tsInformer.GetIndexer().ByIndex("namespace", namespace)
-	if err != nil {
-		log.Errorf("Unable to get list of TransportServers for namespace '%v': %v",
-			namespace, err)
-		return nil
+	var orderedTSs []interface{}
+	var err error
+
+	if namespace == "" {
+		orderedTSs = crInf.tsInformer.GetIndexer().List()
+	} else {
+		orderedTSs, err = crInf.tsInformer.GetIndexer().ByIndex("namespace", namespace)
+		if err != nil {
+			log.Errorf("Unable to get list of TransportServers for namespace '%v': %v",
+				namespace, err)
+			return nil
+		}
 	}
-
-	for _, obj := range orderedVSs {
+	for _, obj := range orderedTSs {
 		vs := obj.(*cisapiv1.TransportServer)
-		// TODO
-		// Validate the TransportServers List to check if all the vs are valid.
-
+		// TODO Validate the TransportServers List to check if all the vs are valid.
 		allVirtuals = append(allVirtuals, vs)
 	}
 
@@ -1320,7 +1320,7 @@ func filterTransportServersForService(allVirtuals []*cisapiv1.TransportServer,
 	return result
 }
 
-func (crMgr *CRManager) getAllServicesFromAllNamespaces() []*v1.Service {
+func (crMgr *CRManager) getAllServicesFromMonitoredNamespaces() []*v1.Service {
 	var svcList []*v1.Service
 	if crMgr.watchingAllNamespaces() {
 		objList := crMgr.crInformers[""].svcInformer.GetIndexer().List()
@@ -1330,7 +1330,7 @@ func (crMgr *CRManager) getAllServicesFromAllNamespaces() []*v1.Service {
 		return svcList
 	}
 
-	for ns, _ := range crMgr.namespaces {
+	for ns := range crMgr.namespaces {
 		objList := crMgr.crInformers[ns].svcInformer.GetIndexer().List()
 		for _, obj := range objList {
 			svcList = append(svcList, obj.(*v1.Service))
@@ -1344,7 +1344,7 @@ func (crMgr *CRManager) getAllServicesFromAllNamespaces() []*v1.Service {
 func (crMgr *CRManager) getVirtualServersForIPAM(ipam *ficV1.F5IPAM) []*cisapiv1.VirtualServer {
 	log.Debug("[ipam] sync ipam starting...")
 	var allVS, vss []*cisapiv1.VirtualServer
-	allVS = crMgr.getAllVSFromAllNamespaces()
+	allVS = crMgr.getAllVSFromMonitoredNamespaces()
 	for _, status := range ipam.Status.IPStatus {
 		for _, vs := range allVS {
 			if status.Host == vs.Spec.Host {
@@ -1359,7 +1359,7 @@ func (crMgr *CRManager) getVirtualServersForIPAM(ipam *ficV1.F5IPAM) []*cisapiv1
 // Get List of TransportServers associated with the IPAM resource
 func (crMgr *CRManager) getTransportServersForIPAM(ipam *ficV1.F5IPAM) []*cisapiv1.TransportServer {
 	var allTS, tss []*cisapiv1.TransportServer
-	allTS = crMgr.getAllTSFromAllNamespaces()
+	allTS = crMgr.getAllTSFromMonitoredNamespaces()
 	for _, status := range ipam.Status.IPStatus {
 		for _, ts := range allTS {
 			key := ts.ObjectMeta.Namespace + "/" + ts.ObjectMeta.Name + "_ts"
@@ -1374,7 +1374,7 @@ func (crMgr *CRManager) getTransportServersForIPAM(ipam *ficV1.F5IPAM) []*cisapi
 
 func (crMgr *CRManager) syncAndGetServicesForIPAM(ipam *ficV1.F5IPAM) []*v1.Service {
 
-	allServices := crMgr.getAllServicesFromAllNamespaces()
+	allServices := crMgr.getAllServicesFromMonitoredNamespaces()
 	if allServices == nil {
 		return nil
 	}
@@ -1586,7 +1586,7 @@ func (crMgr *CRManager) processIngressLink(
 
 	if isILDeleted {
 		var delRes []string
-		for k, _ := range crMgr.resources.rsMap {
+		for k := range crMgr.resources.rsMap {
 			rsName := "ingress_link_" + formatVirtualServerName(
 				ingLink.Spec.VirtualServerAddress,
 				0,